requirements.txt
asgiref==3.3.1
Django==3.1.6
django-extensions==3.1.1
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
Markdown==3.3.3
numpy==1.20.1
pandas==1.2.1
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
